+++ packages/notificare-in-app-messaging/src/internal/iam-component.ts
@@ -1,5 +1,6 @@
 import { Component, getCurrentDevice, getOptions } from '@notificare/web-core';
 import { evaluateContext, handleDocumentVisibilityChanged } from './internal-api';
+import { dismissMessage } from './ui/message-presenter';
 
 /* eslint-disable class-methods-use-this */
 export class IamComponent extends Component {
@@ -24,5 +25,12 @@ export class IamComponent extends Component {
 
   async unlaunch(): Promise<void> {
     //
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  processBroadcast(event: string, data?: unknown) {
+    if (event === 'notification_opened') {
+      dismissMessage();
+    }
   }
 }